+++ challenges/how-to-capitalize-a-string/capitalize.py
