+++ challenges/how-to-check-if-a-key-exists-in-a-dictionary/case_insensitive_lookup.py
