+++ challenges/how-to-check-if-a-key-exists-in-a-dictionary/test_case_insensitive_lookup.py
@@ -1,4 +1,4 @@
-import solutions
+import case_insensitive_lookup
 from inspect import getmembers, isfunction
 
 
@@ -6,5 +6,5 @@ def test_case_insensitive_lookup_capital():
     dictionary = {
         "shrub": "a woody plant which is smaller than a tree and has several main stems arising at or near the ground."
     }
-    for _, value in getmembers(solutions, isfunction):
+    for _, value in getmembers(case_insensitive_lookup, isfunction):
         assert value(dictionary, "Shrub") == dictionary["shrub"]
